Мелкие доработки

+++ app/src/main/java/com/example/falin/simpleweather/Model/CurrentWeather/Coord.kt
@@ -2,5 +2,5 @@ package com.example.falin.simpleweather.Model.CurrentWeather
 
 data class Coord(
         val lon: Double,
-        val lat: Int
+        val lat: Double
 )
